apps/seed-planter-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.1.0
python-multipart==0.0.12
websockets==13.1
aiofiles==24.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
anthropic==0.40.0
PyGithub==2.5.0
GitPython==3.1.43
google-cloud-resource-manager==1.12.5
google-cloud-run==0.10.7
google-auth==2.35.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter for Cloud SQL

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Payment Processing
stripe==7.8.0

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
